+++ cli/config/testdata/bad.01.sq.yml
@@ -1,2 +1,2 @@
-defaults:
+options:
   ping_timeout: not_a_duration
