+++ cli/config/testdata/bad.02.sq.yml
@@ -1,2 +1,2 @@
-defaults:
+options:
   output_format: not_a_format
