+++ cli/config/testdata/bad.03.sq.yml
@@ -1,2 +1,2 @@
-defaults:
+options:
   output_header: not_a_bool
