+++ cli/config/testdata/bad.08.sq.yml
@@ -1,2 +1,2 @@
-defaults:
+options:
   shell_completion_timeout: not_a_duration
