+++ cli/config/testdata/good.02.sq.yml
@@ -1,4 +1,4 @@
-defaults:
+options:
 
 sources:
   active: ""
