+++ cli/config/testdata/good.04.sq.yml
@@ -1,4 +1,4 @@
-defaults:
+options:
 
 sources:
