+++ cli/output/yamlw/mdwriter.go
@@ -1,0 +1,42 @@
+package yamlw
+
+import (
+	"io"
+
+	"github.com/goccy/go-yaml/printer"
+
+	"github.com/neilotoole/sq/cli/output"
+	"github.com/neilotoole/sq/libsq/driver"
+	"github.com/neilotoole/sq/libsq/source"
+)
+
+// mdWriter implements output.MetadataWriter for YAML.
+type mdWriter struct {
+	out io.Writer
+	fm  *output.Formatting
+	p   printer.Printer
+}
+
+// NewMetadataWriter returns a new output.MetadataWriter instance
+// that outputs metadata in JSON.
+func NewMetadataWriter(out io.Writer, fm *output.Formatting) output.MetadataWriter {
+	return &mdWriter{out: out, fm: fm, p: newPrinter(fm)}
+}
+
+// DriverMetadata implements output.MetadataWriter.
+func (w *mdWriter) DriverMetadata(md []driver.Metadata) error {
+	return writeYAML(w.p, w.out, md)
+}
+
+// TableMetadata implements output.MetadataWriter.
+func (w *mdWriter) TableMetadata(md *source.TableMetadata) error {
+	return writeYAML(w.p, w.out, md)
+}
+
+// SourceMetadata implements output.MetadataWriter.
+func (w *mdWriter) SourceMetadata(md *source.Metadata) error {
+	md2 := *md // Shallow copy is fine
+	md2.Location = source.RedactLocation(md2.Location)
+
+	return writeYAML(w.p, w.out, &md2)
+}
